+++ src/qa4sm_reader/globals.py
@@ -80,6 +80,8 @@ map_water_color = "#e0f7fa"
 
 # === boxplot_basic defaults ===
 metadata_max_boxes = 90
+aggregated_max_boxes = 30
+comparison_max_boxes = 90
 metadata_min_samples = 5
 
 boxplot_height = 7 #$ increased by 1 to house logo
@@ -102,8 +104,6 @@ boxplot_width_vertical = 5
 
 period_bin_th = 6 # Above the threshold the figure height gets changes dynamically relative to # of bins
 
-bp_height_exponent = 3/4 # exponent determining how the (non-metadata-)boxplot grows with increasing axes rows
-
 boxplot_edgecolor = "#000000" # color of the edgeline of the boxplot
 boxplot_edgewidth = 1 # width of edgeline of boxplot
 cap_factor = 2/3
@@ -112,6 +112,8 @@ boxplot_new_coloring = True
 hatch_linewidth = 1.5
 num_hatches = 30
 boxplot_printnumbers = True  # Print 'median', 'nObs', 'stdDev' to the boxplot_basic.
+
+bp_height_exponent = 3/4 # exponent determining how the (non-metadata-)boxplot grows with increasing axes rows
 
 #TODO: remove eventually, as watermarlk string no longer needed
 # === watermark defaults ===
@@ -205,10 +207,9 @@ status_replace = {
 def get_status_colors():
     # function to get custom cmap for calculation errors
     # limited to 14 different error entries to produce distinct colors
-    cmap = cl.ListedColormap(matplotlib.colormaps['Set3'].colors[:len(status) - 2])
-    colors = [cmap(i) for i in range(cmap.N)]
-    colors.insert(0, (0, 0.66666667, 0.89019608, 1.0))
-    colors.insert(0, (0.45882353, 0.08235294, 0.11764706, 1.0))
+    colors = list(colorcet.m_CET_L18(np.linspace(0.25,1,len(status)-2)))
+    colors.insert(0, (188/256, 206/256, 217/256, 1.0)) # #bcced9
+    colors.insert(0, (40/256, 0, 0, 1.0))
     cmap = cl.ListedColormap(colors=colors)
     return cmap