+++ src/qa4sm_reader/plotter.py
@@ -34,22 +34,6 @@ matplotlib.rcParams['boxplot.whiskerprops.linewidth'] = globals.boxplot_edgewidt
 matplotlib.rcParams['boxplot.capprops.linewidth'] = globals.boxplot_edgewidth
 matplotlib.rcParams['boxplot.medianprops.linewidth'] = globals.boxplot_edgewidth
 
-_old_boxplot = sns.boxplot
-
-def custom_boxplot(*args, **kwargs):
-    defaults = dict(
-        boxprops=dict(edgecolor=globals.boxplot_edgecolor, linewidth=globals.boxplot_edgewidth),
-        whiskerprops=dict(color=globals.boxplot_edgecolor, linewidth=globals.boxplot_edgewidth),
-        capprops=dict(color=globals.boxplot_edgecolor, linewidth=globals.boxplot_edgewidth),
-        medianprops=dict(color=globals.boxplot_edgecolor, linewidth=globals.boxplot_edgewidth),
-    )
-    for k, v in defaults.items():
-        if k in kwargs:
-            defaults[k].update(kwargs.pop(k))
-    return _old_boxplot(*args, **kwargs, **defaults)
-
-sns.boxplot = custom_boxplot
-
 class QA4SMPlotter:
     """
     Class to create image files of plots from the validation results in a QA4SMImage
@@ -385,242 +369,6 @@ class QA4SMPlotter:
                 continue
 
             yield df, Var, ci
-
-    @staticmethod
-    def _get_dataset_dict(Var) -> dict:
-        """
-        Creates dict containing dataset ids as keys and pretty name (version) as values.
-        version only gets appended if there are multiple datasets witth the same pretty name.
-        
-        Parameters
-        ----------
-        Var : QA4SMMetricVariable
-            Var in the image to make the map for.
-
-        Returns
-        -------
-        d : dict
-            Dict containing id:f"{pretty_name+(version)}" key-value pairs.
-        """
-        dataset_ref = {Var.Datasets.ref_id:f"{Var.Datasets.ref["pretty_name"]}"}
-        dataset_others = {Var.Datasets.others_id[i]:f"{Var.Datasets.others[i]["pretty_name"]}" for i in range(len(Var.Datasets.others))}
-        d = dataset_ref | dataset_others
-        # Append version number if there are multiple datasets with the same pretty name
-        groups = {}
-        for k, v in (d).items():
-            groups.setdefault(v, []).append(k)
-
-        for i in groups.keys():
-            if len(groups[i])>1: 
-                for j in groups[i]:
-                    d[j] = d[j]+f" ({Var.Datasets.dataset_metadata(j)[1]["pretty_version"]})"
-        return d
-    
-    @staticmethod
-    def _get_legend_title(Var) -> str:
-        """
-        Creates a title for an existing legend from a QA4SMMetricVariable.
-
-        Parameters
-        ----------
-        Var : QA4SMMetricVariable
-            Var in the image to make the map for.
-
-        Returns
-        -------
-        legend_title : String
-            String containing the legend title
-        """
-        _, _, _, scale_ds, _ = Var.get_varmeta()
-        d = QA4SMPlotter._get_dataset_dict(Var)
-        
-        # Append Unit
-        for k in d.keys():
-            d[k] = d[k]+f" [{Var.Datasets.dataset_metadata(k)[1]['mu'] if not scale_ds else scale_ds[1]["mu"]}]"
-        
-        legend_title = "Datasets:\n" + "\n".join(f"{k}: {v}" for k, v in (d).items())
-        return legend_title
-    
-    @staticmethod
-    def _append_legend_title(fig, ax, Var) -> tuple:
-        """
-        Appends a title to an existing legend from a QA4SMMetricVariable.
-
-        Parameters
-        ----------
-        fig : matplotlib.figure.Figure
-            The figure that contains the axes and legend.
-        ax : matplotlib.axes.Axes
-            The subplot axis containing the legend to modify.
-        Var : QA4SMMetricVariable
-            Variable object used to construct the legend title.
-
-        Returns
-        -------
-        fig, ax : tuple
-            The same figure and axis, with the legend title updated.
-        """  
-        legend = ax.get_legend()
-        legend_title = QA4SMPlotter._get_legend_title(Var)
-
-        # Change Size to same as rest of legend
-        if len(legend.legend_handles) == 0:
-            legend.set_title(legend_title, prop={'size': globals.fontsize_legend})
-        else:
-            fs = legend.get_texts()[0].get_fontsize()
-            legend.set_title(legend_title, prop={'size': fs})
-
-        legend._legend_box.align = "left"
-        best_loc_with_title = plm.best_legend_pos_exclude_list(ax)
-
-        # Step 4: move legend to new best position
-        legend.set_loc(best_loc_with_title)
-
-        return fig, ax
-    
-    @staticmethod
-    def _smart_suptitle(fig, pad=globals.fontsize_title/2):
-        """
-        Compute position of Suptitle centeredd above axes.
-
-        Parameters
-        ----------
-        fig : matplotlib.figure.Figure
-            The figure object.
-        pad : float
-            Extra space (in fontsize) above the top axes title.
-        """
-        fig.canvas.draw() 
-
-        top_positions = []
-        for ax in fig.axes:
-            if ax.get_visible():
-                # get bounding box of the title in figure coordinates
-                title = ax.title
-                bbox = title.get_window_extent(renderer=fig.canvas.get_renderer())
-                bbox_fig = bbox.transformed(fig.transFigure.inverted())
-                top_positions.append(bbox_fig.y1 + title.get_fontsize()/(72*fig.get_figheight()))
-
-        if top_positions:
-            y = max(top_positions) + pad/(72*fig.get_figheight())
-            y = min(y, 0.99) # So Suptitle always in Figure
-        else:
-            y = 0.99  # fallback if no axes
-        # get center of axes or average of centers if multiple axes
-        x = np.mean([(ax.get_position().x0+ax.get_position().x1)/2 for ax in fig.get_axes()[:globals.n_col_agg]])
-
-        return x, y
-    
-    @staticmethod
-    def _smart_suplabel(fig, axis, pad=globals.fontsize_label/2):
-        """
-        Compute position of suplabels centered according to axes.
-
-        Parameters
-        ----------
-        fig : matplotlib.figure.Figure
-            The figure object.
-        axis : str
-            Axis for which to calculate position.
-        pad : float
-            Extra space (in fontsize) above the top axes title.
-        """
-        fig.canvas.draw() 
-        if axis == "x":
-            bottom_positions = []
-            for ax in fig.axes:
-                if ax.get_visible():
-                    renderer = fig.canvas.get_renderer()
-
-                    # consider x-axis label and tick labels
-                    xlabel_bbox = ax.xaxis.label.get_window_extent(renderer=renderer)
-                    xtick_bboxes = [t.get_window_extent(renderer=renderer) for t in ax.xaxis.get_ticklabels() if t.get_text()]
-                    
-                    all_bboxes = [xlabel_bbox] + xtick_bboxes
-                    all_bboxes_fig = [b.transformed(fig.transFigure.inverted()) for b in all_bboxes]
-                    bottom_positions.append(min(b.y0 for b in all_bboxes_fig))
-
-            if bottom_positions:
-                y = min(bottom_positions) - globals.fontsize_label/(72*fig.get_figheight()) - pad/(72*fig.get_figheight())
-            else:
-                y = 0.01  # fallback
-            x = np.mean([(ax.get_position().x0+ax.get_position().x1)/2 for ax in fig.get_axes()[:globals.n_col_agg]])
-            return x, y
-        elif axis == "y":
-            left_positions = []
-            for ax in fig.axes:
-                if ax.get_visible():
-                    renderer = fig.canvas.get_renderer()
-
-                    # consider x-axis label and tick labels
-                    ylabel_bbox = ax.yaxis.label.get_window_extent(renderer=renderer)
-                    ytick_bboxes = [t.get_window_extent(renderer=renderer) for t in ax.yaxis.get_ticklabels() if t.get_text()]
-                    
-                    all_bboxes = [ylabel_bbox] + ytick_bboxes
-                    all_bboxes_fig = [b.transformed(fig.transFigure.inverted()) for b in all_bboxes]
-                    left_positions.append(min(b.x0 for b in all_bboxes_fig))
-
-            if left_positions:
-                x = min(left_positions) - globals.fontsize_label/(72*fig.get_figwidth()) - pad/(72*fig.get_figheight()) 
-            else:
-                x = 0.01  # fallback
-            y = np.mean([(ax.get_position().y0+ax.get_position().y1)/2 for ax in fig.get_axes()[::globals.n_col_agg]])
-            return x, y
-        else: #fallback
-            return 0.01, 0.01
-
-    def _get_ax_width(fig) -> float:
-        """Get horizontal distance of all axes in px. From left of first in row to right of last in row."""
-        left = min([ax.get_position().x0 for ax in fig.get_axes()[:1]]) # Always the first ax
-        right = max([ax.get_position().x1 for ax in fig.get_axes()])
-        ax_width_px = fig.get_figwidth()*(right-left) * fig.dpi
-
-        return ax_width_px
-    
-    def _get_ax_height(fig) -> float:
-        """Get vertical distance of all axes in px. From bottom of column to top of column."""
-        bottom = min([ax.get_position().y0 for ax in fig.get_axes()]) # Always the first ax
-        top = max([ax.get_position().y1 for ax in fig.get_axes()])
-        ax_height_px = fig.get_figheight()*(top-bottom) * fig.dpi
-
-        return ax_height_px
-
-    @staticmethod
-    def _set_wrapped_title(fig, ax, title, fontsize=globals.fontsize_title,
-                           pad=globals.title_pad, use_suptitle=False):
-        """
-        Set an axes or figure suptitle that automatically wraps to fit within figure width.
-        
-        Parameters
-        ----------
-        fig : matplotlib.figure.Figure
-            The figure object.
-        ax : matplotlib.axes.Axes
-            The target axes (ignored if use_suptitle=True).
-        title : str
-            The title text.
-        fontsize : int
-            Font size of the title.
-        pad : float
-            Extra spacing from the plot, in points.
-        use_suptitle : bool, optional
-            If True, set a figure-wide suptitle instead of an axes title.
-        """
-        fig.canvas.draw()  # needed to get renderer
-
-        # width between most left point and most right point in axesin inches * dpi = pixels
-        ax_width_px = QA4SMPlotter._get_ax_width(fig)
-
-        # estimate character width (rough, depends on font)
-        wrapped = th.wrapped_text(fig, title, ax_width_px, fontsize)
-
-        if use_suptitle:
-            x, y = th.smart_suptitle(fig)
-            fig.suptitle(wrapped, fontsize=fontsize, y=y, x=x, ha="center", va="bottom")
-
-        else:
-            x, y = th.smart_suptitle(fig)
-            ax.set_title(wrapped, fontsize=fontsize, pad=pad)
 
     def _boxplot_definition(self,
                             metric: str,
@@ -1018,6 +766,7 @@ class QA4SMPlotter:
         if scl_meta:
             scl_short = scl_meta[1]['short_name']
 
+        is_scattered = Var.attrs.get('val_is_scattered_data') == 'True'
         # create mapplot
         fig, ax = plm.mapplot(
             df=Var.values[Var.varname],
@@ -1027,6 +776,7 @@ class QA4SMPlotter:
             plot_extent=
             None,  # if None, extent is automatically adjusted (as opposed to img.extent)
             scl_short=scl_short,
+            is_scattered=is_scattered,
             **style_kwargs)
 
         # title and plot settings depend on the metric group
@@ -1324,7 +1074,8 @@ class QA4SMPlotter:
         else:
             ax_first = axes.flat[0]
 
-        ax_first.legend(fontsize=globals.fontsize_legend).set_loc("upper left")
+        ax_first.legend(fontsize=globals.fontsize_legend, 
+                        ncol=(len(ax_first.get_legend_handles_labels()[0])-1)//5 + 1).set_loc("upper left")
         _, ax_first = th.append_legend_title(fig, ax_first, Var)
         
         return fig, axes
@@ -1528,6 +1279,9 @@ class QA4SMPlotter:
                     warnings.warn(
                         f"Too many boxes would be drawn in '{meta_type}` "
                         f"metadata-based `{metric}` plots. Neglected due to overcrowding")
+                elif "no valid measurement" in msg:
+                    warnings.warn(f"There are no valid measurements for this metric ({metric}) metadata ({meta_type}) combination "
+                         "therefore the creation of a plot is skipped.")
                 else:
                     warnings.warn(
                         f"'{meta_type}` metadata-based `{metric}` plot wasn't created.")
@@ -2118,13 +1872,17 @@ class QA4SMCompPlotter:
         df_long = metric_df.stack(level=0, future_stack=True)
         df_long.index = df_long.index.set_names(["lat", "lon", "gpi", "dataset"])
         df_long = df_long.reset_index()
-        df_long = df_long.melt(id_vars=["lat", "lon", "gpi", "dataset"], var_name="year", value_name="value").sort_values("dataset", ascending=True)
+        df_long = df_long.melt(id_vars=["lat", "lon", "gpi", "dataset"], var_name="year", value_name="value").sort_values(["dataset", "year"], ascending=True)
         
         unique_combos = df_long["dataset"].unique()
         palette = get_palette_for(unique_combos)
 
         n_lines = len(cbp_fig.ax_box.lines)
-        box = sns.boxplot(x="year",
+        if len(df_long["value"]) == int(df_long["value"].isna().sum()):
+            raise PlotterError(f"There are no valid measurements for this metric ({chosen_metric}) "
+                            "therefore the creation of a plot is skipped.")
+    
+        box = plm.sns_custom_boxplot(x="year",
                               y="value",
                               hue="dataset",
                               data=df_long,
@@ -2137,7 +1895,16 @@ class QA4SMCompPlotter:
         plm.capsizing(box, n_lines)
             
         box.tick_params(labelsize=globals.fontsize_ticklabel)
-        cbp_fig.ax_box.legend(loc=plm.best_legend_pos_exclude_list(cbp_fig.ax_box), fontsize=globals.fontsize_legend)
+
+        # Check if duplicate legend entries
+        handles, labels = cbp_fig.ax_box.get_legend_handles_labels()
+        unique = dict(zip(labels, handles))
+
+        cbp_fig.ax_box.legend(unique.values(),
+                              unique.keys(),
+                              loc=th.best_legend_pos_exclude_list(cbp_fig.ax_box), 
+                              fontsize=globals.fontsize_legend, 
+                              ncol=(len(unique)-1)//5 + 1)
         
         Var = hdl.MetricVariable(varname=self.metrics_ds_grouped_lut()[chosen_metric][list(self.metrics_ds_grouped_lut()[chosen_metric].keys())[0]], global_attrs=self.ds.attrs)
         th.append_legend_title(cbp_fig, cbp_fig.ax_box, Var)
@@ -2201,4 +1968,4 @@ class QA4SMCompPlotter:
         ]
 
         return cbp_fig.fig
-
+