+++ CHANGELOG.rst
@@ -4,19 +4,34 @@ Changelog
 
 Unreleased changes in master branch
 ===================================
+- Pin netcdf4 to <1.7.3
+
+Version 0.11.6
+==============
+- Bugfix creation of comparison mapplot
+
+Version 0.11.5
+==============
+- Update build.yml
+- Fix comparison module
+- Results with val_is_scattered_data == 'True' rendered as point data in plots
+
+Version 0.11.4
+==============
+- Added Custom Plot Functions for Users support materials
 
 Version 0.11.3
-===================================
+==============
 - Added fallback names for new CCI and C3S data
 - Pyscaffold version updated, use pyproject.toml instead of setup.cfg
 
 Version 0.11.2
-===================================
+==============
 - Update dependencies, remove environment.yml
 - Fix empty comparison plot generation in stability case
 
 Version 0.11.1
-===================================
+==============
 - ISMN metadata boxplots split by network (`PR #97 <https://github.com/awst-austria/qa4sm-reader/pull/97>`_)
 - Fix filenames of triple collocation mapplots
 - Restrict qtbase to versions <= 6.7.2